Admin Generator: Refactor asyncSelect generation into own file
